// my-profile/src/enums/E_criterion.ts
enum E_criterion {
  Extraversion = "Extraversion", // 선호하는 세계 ( E )
  Introversion = "Introversion", // 선호하는 세계 ( I )
  Intuition = "Intuition", // 인식형태 ( N )
  Sensing = "Sensing", // 인식형태 ( S )
  Thinking = "Thinking", // 판단기준 ( T )
  Feeling = "Feeling", // 판단기준 ( F )
  Perceiving = "Perceiving", // 생활양식 ( p )
  Judging = "Judging", // 생활양식 ( J )
}
export default E_criterion;